requirements.txt
fastapi
uvicorn[standard]
aiohttp
beautifulsoup4
openai
pandas
matplotlib
scikit-learn
google-generativeai
